# docs/swagger/openapi3_def.yaml
components:
  parameters:
    auth:
      allowEmptyValue: false
      description: Authorization Headers.
      in: header
      name: auth
      required: true
      schema:
        type: string
...
    packageName:
      allowEmptyValue: false
      description: The name of the package to return details for. Must be URL escaped.
      example: autocomplete-powershell
      in: path
      name: packageName
      required: true
      schema:
        type: string
...
    rename:
      allowEmptyValue: false
      description: The new package name to rename to, if applicable.
      example: new-package-name
      in: query
      name: rename
      required: false
      schema:
        type: string
...
  schemas:
...
    packageObjectFull:
      description: A 'Package Object Full' of a package on the PPR.
      properties:
        badges:
          type: array
        creation_method:
          type: string
        downloads:
          type: string
        metadata:
          type: object
        name:
          type: string
        owner:
          type: string
        readme:
          type: string
        releases:
          type: object
        repository:
          type: object
        stargazers_count:
          type: string
        versions:
          type: object
      required:
        - name
        - readme
        - metadata
        - releases
        - versions
        - repository
        - creation_method
        - downloads
        - stargazers_count
        - badges
        - owner
      type: object
...
info:
  description: >-
    Allows for the management, viewing, and downloading of packages and themes
    for use within the Pulsar editor.
  license:
    identifier: MIT
    name: MIT
  title: Pulsar
  version: 1.1.0
openapi: 3.1.0
paths:
...
  /api/packages/{packageName}/versions:
    options:
      responses:
        '204':
          description: The options response for this endpoint.
          headers:
            Allow:
              description: The allowed HTTP methods on this endpoint.
              schema:
                type: string
            X-Content-Type-Options:
              description: Indicates the 'Content-Type' header should be followed.
              schema:
                type: string
    post:
      parameters:
        - $ref: '#/components/parameters/rename'
        - $ref: '#/components/parameters/auth'
        - $ref: '#/components/parameters/packageName'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/packageObjectFull'
          description: The package details indicating success.
      summary: Creates a new package version.
...
  /api/themes/{packageName}/versions:
    options:
      responses:
        '204':
          description: The options response for this endpoint.
          headers:
            Allow:
              description: The allowed HTTP methods on this endpoint.
              schema:
                type: string
            X-Content-Type-Options:
              description: Indicates the 'Content-Type' header should be followed.
              schema:
                type: string
    post:
      parameters:
        - $ref: '#/components/parameters/rename'
        - $ref: '#/components/parameters/auth'
        - $ref: '#/components/parameters/packageName'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/packageObjectFull'
          description: The package details indicating success.
      summary: Creates a new package version.
...
servers:
  - description: Production Server
    url: https://api.pulsar-edit.dev
  - description: Locally hosted development server
    url: http://localhost:8080